[python] Rework of enum handling to simplify and add features [behavior change]
So now, if you	 have an enum:

    <enum name="CarColor" encodingType="char">
        <validValue name="Black">B</validValue>
        <validValue name="White">W</validValue>
    </enum>

The generated CarColor now has a __str__() method for displaying the name of the enum value.

When retrieving the value of an enum field, you now get an object that is an instance of
the enum.  So:

    print('The color is: {}'.format(car.getColor()))

would display something like: "The color is: Black" rather than "The color is: 66"
or "The color is: B".

To encode an enum, you would pass a value of the enum like:

    car.setColor(CarColor.Black)

or the raw value:

    car.setColor('B')

Since retrieving an emum yields an instance of the enum, its underlying value is in
its "value" property:

    if car.getColor().value == CarColor.Black:
        handleBlackCar(car)