Add the golang keywords to the set of identifier checks.
This causes the Car example to change as 'make' is reserved which is
changed to 'brand'. This has cascading effects into the samples and
benchmarks for both Java and CPP but it remains sane.

It also adjust the number of expected errors in two of the
ErrorHandler tests as 'true' and 'false' get additionally counted for
golang.