[golang] Issue #435 References inside messageHeader not working
Generate the MessageHeader as though it was a composite as it's not compulsorarily standardized in the standard.

Provide an optimized separate MessageHeader for use when people follow the recommended MessageHeader as most people will do just that.

Fix <ref tags> to use the underlying type find token.acceptableTypeName() rather than token.name(). In golang references had previously been worked around for sets/choices/composites but not MessageHeaders. This makes them all work the same way and more properly but it does introduce some monor changes to generated types.

Fix the tests, examples and benchmarks to reflect these changes, adding an issue435.xml for testing with test code to match.

Fix the java interop test which used the sbe-samples example-schema which has now been modified. This also causes the bigendian test to be revised.

Documentation updates to match.