[Go] Fix type name for nested composites (#847)
There was a typo in type name generation of nested
composites.  It was using the `propertyName` instead
of `propertyType`.

Add a test case and unit test. This situation wasn't
elicited before because nested property types and names
are the same in tests (e.g. `Engine`).

Prior to the change, the test case in this changeset would show:
`./MessageHeader.go:16:14: undefined: MessageHeaderC1`

Also set GO111MODULE in Makefile for Go 1.16 builds.