[Go] fix excess import generation (#848)
The previous implementation used a single `TreeSet` member to track the import
names.  This was done as a member to not have to pass the imports to many
functions.

A given file should only have its imports.  Due to traversal of messages and
components, there are times where there are too many.  The `TreeSet` needs to
be fresh for each file.

This commit changes `imports` from a single `TreeSet` to a `Stack<TreeSet>`.

Each time where there was a line:
`imports = new TreeSet<>()`  (always near a `outputManager.createOutput` new file)

We now push that TreeSet on the stack:
`imports.push(new TreeSet<>())`

At the end of the block, that is popped.  This keeps the imports scoped to
releveant generators.

All those generators now `peek()` at the current `TreeSet`.