[Go] GolangGenerator improvements (#846)
* [Go] Refactor whitespace in GolangGenerator (#844)

Adds function `generateWhitespace` which uses the same String.format
technique, but validates the input.  It also always generates at least
one space.

This change should only affect the whitespace of the output.  It does
not attempt to change the logic that might be the underlying cause of #844.

* [Go] Fix type name for nested composites (#847)

There was a typo in type name generation of nested
composites.  It was using the `propertyName` instead
of `propertyType`.

Add a test case and unit test. This situation wasn't
elicited before because nested property types and names
are the same in tests (e.g. `Engine`).

Prior to the change, the test case in this changeset would show:
`./MessageHeader.go:16:14: undefined: MessageHeaderC1`

Also set GO111MODULE in Makefile for Go 1.16 builds.

* [Go] fix excess import generation (#848)

The previous implementation used a single `TreeSet` member to track the import
names.  This was done as a member to not have to pass the imports to many
functions.

A given file should only have its imports.  Due to traversal of messages and
components, there are times where there are too many.  The `TreeSet` needs to
be fresh for each file.

This commit changes `imports` from a single `TreeSet` to a `Stack<TreeSet>`.

Each time where there was a line:
`imports = new TreeSet<>()`  (always near a `outputManager.createOutput` new file)

We now push that TreeSet on the stack:
`imports.push(new TreeSet<>())`

At the end of the block, that is popped.  This keeps the imports scoped to
releveant generators.

All those generators now `peek()` at the current `TreeSet`.