[Java] Add checking around decoding order.
This change builds upon the work to check field encoding order.
It is a slightly more-complex problem than with encoding, as when
decoding it is necessary to deal with multiple versions of encoded data.
To handle this, we fuse together the state machines for decoding each
possible version and branch on "decoder.wrap(...)".

Further work:

- More tests
- Better error messages