[Java] Test field order checks across versions.
"New" decoders (in Java at least) read null-like values for fields
that are not present in "old" messages. There are no restrictions on the
order in which these "new" fields are accessed when decoding "old" data.

In Java, a version check returns a null-like value when a field is not
present in the acting version. This behaviour matches the wiki
documentation.

C#, currently, has different behaviour: the application code needs to
interrogate a boolean property, e.g., `IsFooInActingVersion`, before
accessing the data property, e.g., `Foo`. The property itself assumes
that the field is present. It will attempt to read it from the buffer
and possibly advance a `limit` field.

In this commit, I have moved the state transition beneath the existing
version check in Java decoders. This change allows these fields to be
accessed when decoding "old" messages without changing the state of
decoding, which seems appropriate as there are no side-effects on
internal state, e.g., the `limit` field.

Also, in this commit, I have expanded the test coverage to include the
getter and setter variants for fixed-size arrays and variable-length
data.