[Java] Treat top-level block fields as a special case.
To reduce the number of false negatives the field access order checker
generates, in this commit, we treat the top-level block fields
differently.

We know that once the top-level encoder is wrapped, these fields are
always safe to access.

As a result, we were able to enable several tests that previously
failed.