[Java] Generate check that encoding is complete.
Previously, we checked the order of encoding was correct but not that
the encoding reached a terminal state.

In this commit, I've added code to generate a new method
`checkEncodingIsComplete` on Java encoders. It verifies that a terminal
state has been reached; otherwise, it throws an exception with
information about the current state and transitions.