[Java, C#, C++] Model `resetCountToIndex()` state machine transition.
Previously, the access order checks did not work with the
`resetCountToIndex()` methods (generated on flyweights for repeating
groups). Now, the valid transitions using these methods are modelled in
the state machine and this is reflected in the code we generate for
Java, C# and C++.

Note that the `resetCountToIndex()` method does not change the
`limit`/`position` of the message; therefore, it is only valid to use it
when the `limit` aligns with a boundary between group elements.