[Java] Extend PBT to generate arbitrary encoded messages.
In this commit, I've also demonstrated the usage of arbitrary encoded
messages in a JsonPrinter test. In the test, I print an arbitrary value
of an arbitrary schema and check that it is valid JSON.

The test passes as long as we constrain the generation of strings, as
JsonPrinter currently does not do any escaping. I also noted that it
uses `'` rather than `"`, which AFAICS is not permitted by the JSON
specification, but is permitted by our (less strict) JSON testing tool.