[C++] Support "to string" without specification of buffer length.
As we use the generated codecs to create a string representation of our
DTOs, we don't use Agrona buffers in C++, and there is no concept of
resizing, it is necessary to size a temporary buffer during the
construction of the string data.

Previously, we were letting the user supply this value, which wasn't a
very friendly API.

Now, we use the `computeLength` methods on the codec to determine how
big of a temporary buffer we need.

Perhaps the methods will also be useful for avoiding a buffer copy when
used in conjunction with Aeron. For example, a developer could use
`dto.computeEncodedLength()` to initialise a buffer claim rather than
copying via the `offer(...)` API.