[C++] Add PBT for C++ DTOs.
The main change in this commit is to add property-based testing around
the C++ DTOs. We check the same property as the C# tests, i.e., that
decoding and re-encoding via a DTO preserves the original bytes.

There are some other smaller changes in this commit:

1. We now only generate arbitrary schemas where enums have at least one
case, as this is required by the spec.

2. We now log details about how we encoded the `input.dat` file used in
property-based tests, which can help diagnose problems in the test.