Address Mike's feedback around depending directly on system properties.
Also, addresses Mike's feedback around the names of those system
properties.

In this commit, I've introduced a factory-like class:
`PrecedenceChecks`. It creates models for message field access when the
checks are enabled. It also provides a convenient place to group
together some properties affecting the generation of field precedence
checks.

I have also renamed `AccessOrderModel` to `FieldPrecedenceModel` in this
commit, as it aligns better with the system property names.