Address Mike's feedback: move sys prop eval higher.
In this commit, I've removed the `PrecedenceChecks#Configuration` class
in favour of a static method within `TargetCodeGeneratorLoader`.

Unfortunately, C# behaves differently to all the other code generators.
To avoid duplicating the default configuration for precedence checks in
the `CSharp` class, I've made the method publicly visible, which is a
bit ugly.

It has also introduced a cyclic dependency between the code generators,
e.g., `JavaGenerator`, and `TargetCodeGeneratorLoader`, as in order to
preserve the existing public signatures I have kept the old constructors
around and supplied a default `PrecedenceChecks`.