[Java] Add support for DTO generation.
This commit adds support for generating SBE DTOs in Java.

Unlike the C# and C++ support, there isn't a natural candidate for
the idiomatic representation of optional fields, e.g., `Optional<T>` is
only meant for use as a return type. Therefore, the mapping between Java
types and the encoding is simpler than in these other languages.

In a future commit, I will extend the property-based tests to cover Java
DTOs.