[Java] Fix remaining DTO issues uncovered with PBT.
The main issue was around determining when a field might not be present.
We were checking `token.version() > 0`, which isn't accurate in
composites or groups.

In this commit, I've also started to use the `Footnotes` API in JQWIK,
which allows useful debug information to be captured _only_ for failing
attempts. It then prints out the shrunken footnotes upon a failure.