[C#] Test package override support compiles.
Co-authored-by: Nate Bradac <nate.bradac@weareadaptive.com>

We've updated the C# generation tasks to include a schema that uses
explicit package names on `types` blocks.

We found processing this schema produces some build errors (albeit not
affecting the exit code):

```
> Task :generateCSharpTestCodecs
[Error] explicit-package-test-schema.xml:17:49: cvc-complex-type.3.2.2: Attribute 'package' is not allowed to appear in element 'types'.
[Error] explicit-package-test-schema.xml:30:37: cvc-complex-type.3.2.2: Attribute 'package' is not allowed to appear in element 'types'.

> Task :generateCSharpTestDtos
[Error] explicit-package-test-schema.xml:17:49: cvc-complex-type.3.2.2: Attribute 'package' is not allowed to appear in element 'types'.
[Error] explicit-package-test-schema.xml:30:37: cvc-complex-type.3.2.2: Attribute 'package' is not allowed to appear in element 'types'.
```

Therefore, we've put the new generation under tasks that don't validate
against an XSD. We considered creating a new version of the XSD that
properly describes what we support; however, the licence on the XSD
seems prohibitive.